编译告警 Signed-off-by:liupeng298<liupeng298@huawei.com>

+++ interfaces/innerkits/faultloggerd_client/faultloggerd_client.cpp
@@ -292,7 +292,7 @@ static bool SendRequestToServer(const FaultLoggerdRequest &request)
             DfxLogError("memset_s failed, err = %d.", (int)ret);
             break;
         }
-        if (read(sockfd, ControlBuffer, sizeof(ControlBuffer) - 1) != (int)strlen(FAULTLOGGER_DAEMON_RESP)) {
+        if (read(sockfd, ControlBuffer, sizeof(ControlBuffer) - 1) != (ssize_t)strlen(FAULTLOGGER_DAEMON_RESP)) {
             break;
         }
